requirements.txt
autopep8==1.7.0
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
diffusers==0.3.0
dotenv-python==0.0.1
filelock==3.8.0
Flask==2.2.2
google-auth==2.11.0
huggingface-hub==0.9.1
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.3
packaging==21.3
Pillow==9.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pyparsing==3.0.9
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
rsa==4.9
six==1.16.0
tokenizers==0.12.1
toml==0.10.2
torch==1.12.1
tqdm==4.64.1
transformers==4.22.0
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
